Switch to using npm for package @voibo/desktop-audio-capture instead of git submodule
Package desktop-audio-capture uses Node addons and was initially managed
as a Git submodule. However, this approach is not standard practice.
Ideally, the package should be managed through npm. Therefore, this
change transitions the package management from a Git submodule to npm.

- Remove package desktop-audio-capture submodule from third_party directory.
- Remove package bindings from npm dependencies.
- Add package @voibo/desktop-audio-capture to dependencies using npm.
- Update existing code to import and use pacakge
  @voibo/desktop-audio-capture.
- Add package @voibo/desktop-audio-capture to externals in webpack.config.ts.
- Remove the configuration in webpack.config.ts for copying native
  modules of desktop-audio-capture.
- Add Node addon binary path for package @viobo/desktop-audio-capture to
  "asarUnpack" in electron-builder build configuration.

This change requires the installation of tools such as cmake and gcc,
which are necessary to build the native code for the Node addon.

Signed-off-by: Ryosuke Akiyama <r@554.jp>